Change some things in vimrc
The first thing was to get rid off tab control. I dont use tabs and dont
have any intention to start using it, so... get off !

The next change was to kill the change splits biding in order to include
the ctrl l ctrl h to walk the line. This change makes my day a bit
harder but... thats the life.

The changes to walk inside a virtually broke line was deleted. I am not
using it ...

Mapped the "-"and  "_" key to a break line auxiliar.

Mapped the c-m c-l and c-h keys to be able to access the M H L  keys.
Fast move around is the key here !

A quick thing is to fix the leader key dependence on timeout. I dont
like the intermodes delay of vim but I need then to be able to use the
leader key. Thats the problem. Keep the leader key and dont have delays.

On branch master
Your branch is up-to-date with 'origin/master'.

Changes to be committed:
    modified:   vimrc